Trim comments and consolidate tests: merge duplicate updater/seamless-update test files, drop redundant flow test
